+++ test/chez/scan-error-app/src/app/core.clj
@@ -1,3 +1,0 @@
-(ns app.core
-  (:require [app.dup]))
-(defn -main [& _] (println (app.dup/f)))